@@ -1,42 +1,42 @@
 {
-  "name": "react-moving-text-demo",
-  "homepage": "https://yidaoj.github.io/react-moving-text",
-  "version": "0.0.0",
+  "name": "react-style-text-demo",
+  "homepage": "https://andrew87e.github.io/",
+  "version": "0.0.1",
   "description": "",
   "main": "index.js",
   "scripts": {
     "build": "webpack",
     "start": "webpack-dev-server --open"
   },
-  "author": "",
+  "author": "andrewe87e",
   "license": "MIT",
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.26",
-    "@fortawesome/free-solid-svg-icons": "^5.12.0",
-    "@fortawesome/react-fontawesome": "^0.1.8",
-    "prettier": "^1.19.1",
-    "prismjs": "^1.19.0",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-moving-text": "0.0.7",
-    "react-pose": "^4.0.10",
-    "react-router-dom": "^5.1.2",
-    "styled-components": "^4.4.1",
-    "webpack": "^4.41.4"
+    "@fortawesome/fontawesome-svg-core": "6.5.1",
+    "@fortawesome/free-solid-svg-icons": "6.5.1",
+    "@fortawesome/react-fontawesome": "0.2.0",
+    "buffer": "^6.0.3",
+    "framer-motion": "^10.16.16",
+    "prismjs": "1.29.0",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-router-dom": "6.21.1",
+    "react-style-text": "^0.1.27",
+    "stream-browserify": "^3.0.0",
+    "styled-components": "6.1.3",
+    "sweetalert": "^2.1.2",
+    "webpack": "^5.89.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.7.7",
-    "@babel/preset-env": "^7.7.7",
-    "@babel/preset-react": "^7.7.4",
-    "babel-loader": "^8.0.6",
-    "babel-plugin-prismjs": "^2.0.0",
-    "css-loader": "^3.4.2",
-    "eslint": "^6.8.0",
-    "eslint-plugin-prettier": "^3.1.2",
-    "html-loader": "^0.5.5",
-    "html-webpack-plugin": "^3.2.0",
-    "style-loader": "^1.1.2",
-    "webpack-cli": "^3.3.10",
-    "webpack-dev-server": "^3.10.1"
+    "@babel/core": "7.23.6",
+    "@babel/preset-env": "7.23.6",
+    "@babel/preset-react": "7.23.3",
+    "babel-loader": "9.1.3",
+    "babel-plugin-prismjs": "2.1.0",
+    "css-loader": "6.8.1",
+    "html-loader": "4.2.0",
+    "html-webpack-plugin": "5.6.0",
+    "style-loader": "3.3.3",
+    "webpack-cli": "5.1.4",
+    "webpack-dev-server": "4.15.1"
   }
 }
