@@ -1,24 +1,56 @@
-import React from 'react'
-import styled from 'styled-components'
-import MovingComponent from 'react-moving-text'
-import ExternalLink from '../atoms/externalLink'
+import React from "react";
+import styled from "styled-components";
+import { StyledText } from "react-style-text";
+import ExternalLink from "../atoms/externalLink";
 
 const Footer = () => {
+  const me = "andrew87e";
+  const itsaMe = me.split("");
+
   return (
     <FooterContainer>
       Made with
-      <MovingComponent
-        type="flipHorizontal"
-        iteration="infinite"
-        duration="2000ms"
-        timing="linear"
-        style={{margin: "0px 10px", color:"orange"}}>
+      <StyledText
+        style={{ margin: "0px 10px", color: "red" }}
+        animationProps={{
+          animationname: "spin",
+          duration: "2000ms",
+          delay: "0ms",
+          direction: "normal",
+          timing: "linear",
+          iteration: "infinite",
+          fillmode: "none",
+        }}
+      >
         ❤
-      </MovingComponent>
-       by <ExternalLink href="https://github.com/YiDaoJ" target="_blank" margin="0px 10px" color="#2196f3" >YiDaoJ</ExternalLink> using React and styled-components
+      </StyledText>
+      by&nbsp;
+      {itsaMe.map((letter, index) => (
+        <StyledText
+          key={index}
+          animationProps={{
+            animationname: "flipIn",
+            duration: "2000ms",
+            delay: ` ${index * 100}ms`,
+            direction: "normal",
+            timing: "ease-in-out",
+            iteration: "1",
+            fillmode: "none",
+          }}
+        >
+          <ExternalLink
+            href="https://github.com/andrew87e"
+            target="_blank"
+            color="#5cbd6b"
+          >
+            {letter}
+          </ExternalLink>
+        </StyledText>
+      ))}
+      &nbsp; using React and styled-components 💅🏽
     </FooterContainer>
-  )
-}
+  );
+};
 
 const FooterContainer = styled.div`
   display: flex;
@@ -28,9 +60,10 @@ const FooterContainer = styled.div`
   bottom: 0;
   width: 100vw;
   height: 40px;
-  font-size: .8rem;
-  color: ${props => props.theme.colors.middle};
-  background-color: ${props => props.theme.colors.secondary};
-`
+  font-size: 0.8rem;
+  color: ${(props) => props.theme.colors.middle};
+  // background-color: ${(props) => props.theme.colors.secondary};
+  background: radial-gradient(at left top, #000, #01030D);
+`;
 
-export default Footer
+export default Footer;
