@@ -1,12 +1,15 @@
-import React from 'react'
+import React, {useRef} from 'react'
 import styled from 'styled-components'
 import StyledLink from '../atoms/link'
 import ExternalLink from '../atoms/externalLink'
+import { StyledText } from 'react-style-text'
+
 
 const Header = () => {
+  const textAreaRef = useRef(null);
   return (
     <NavBar>
-      <ForkRibbon href="https://github.com/YiDaoJ/react-moving-text" target="_blank">
+      <ForkRibbon href="https://github.com/andrew87e/react-style-text" target="_blank">
         <img
           width="149"
           height="149"
@@ -17,23 +20,65 @@ const Header = () => {
         />
       </ForkRibbon>
       <StyledH1>
-        <StyledLink color="#1976d2" noTransition={true} to="/react-moving-text/">
-          React-Moving-Text
+        <StyledText
+          animationProps={{
+            animationname: "fadeInFromTop",
+            duration: "1000ms",
+            delay: "0ms",
+            direction: "normal",
+            timing: "ease-in-out",
+            iteration: "1",
+            fillmode: "none",
+          }}
+        >
+        <StyledLink color="#fff" noTransition={true} to="/">
+          React-Style-Text
         </StyledLink>
+        </StyledText>
       </StyledH1>
+      
+      <StyledText
+          animationProps={{
+            animationname: "fadeIn",
+            duration: "1000ms",
+            delay: "0ms",
+            direction: "normal",
+            timing: "ease-in-out",
+            iteration: "1",
+            fillmode: "none",
+
+          }}
+        >
       <StyledSubHeader>
         A library for animated texts built with React and CSS Animation.
         <LinkGroupContainer>
-          <StyledLink to="/react-moving-text/example" color="#0d47a1" margin="0px 30px">More Examples</StyledLink>
-          <ExternalLink href="https://github.com/YiDaoJ/react-moving-text/blob/master/README.md" target="_blank" color="#0d47a1">Document</ExternalLink>
+          {/* <StyledLink to="/react-style-text/example" color="#0d47a1" margin="0px 30px">More Examples</StyledLink> */}
+          <ExternalLink href="https://github.com/andrew87e/react-style-text/blob/master/README.md" target="_blank" color="#5cbd6b">Documentation</ExternalLink>
         </LinkGroupContainer>
       </StyledSubHeader>
+        </StyledText>
+        <StyledText 
+          animationProps={{
+            animationname: "fadeInFromBottom",
+            duration: "1000ms",
+            delay: "0ms",
+            direction: "normal",
+            timing: "ease-in-out",
+            iteration: "1",
+            fillmode: "none",
+          }}
+        >
       <InstallInfoBar>
-        <code>npm i react-moving-text</code>
+        <StyledCode onClick={() => {navigator.clipboard.writeText(textAreaRef.current.innerText); swal("Copied!", "Code copied to clipboard!", "success", {button: "Aww yiss!"}) }} ref={textAreaRef}>npm i react-style-text</StyledCode>
       </InstallInfoBar>
+        </StyledText>
     </NavBar>
   )
 }
+
+const StyledCode = styled.code`
+text-shadow: 0.5px 0.5px 0px #5cbd6b;
+`
 
 const NavBar = styled.div`
   display: flex;
@@ -43,7 +88,9 @@ const NavBar = styled.div`
   width: 100vw;
   height: 190px;
   padding: 20px 100px;
-  background-color: ${props => props.theme.colors.primary};
+  // background-color: ${props => props.theme.colors.primary};
+  background: radial-gradient(at center bottom, #323234, #01030D)
+
 `
 
 const ForkRibbon = styled.a`
@@ -58,13 +105,12 @@ const StyledH1 = styled.h1`
   font-weight: 300;
   letter-spacing: -2;
   margin: 0;
-  color: ${props => props.theme.colors.middle};
 `
 
 const StyledSubHeader = styled.h2`
   font-weight: normal;
   font-size: 1rem;
-  color: ${props => props.theme.colors.secondary};
+  color: ${props => props.theme.colors.primary}
 `
 const LinkGroupContainer = styled.div`
   display: inline;
@@ -79,6 +125,8 @@ const InstallInfoBar = styled.div`
   border-radius: 0;
   font-size: 1rem;
   padding: 12px 25px;
+  cursor: pointer;
+  box-shadow: inset 10px 10px 10px rgba(0,0,0,0.5);
 `
 
 export default Header
