@@ -1,5 +1,6 @@
-import React from 'react'
+import React, {useRef} from 'react'
 import styled from 'styled-components'
+import swal from 'sweetalert'
 
 const CodeContainer = ({
   animationObject,
@@ -9,54 +10,59 @@ const CodeContainer = ({
   direction,
   timingFunction,
   iterationCount,
-  fillMode,
+  fillmode,
   interval
 }) => {
+  const textAreaRef = useRef(null);
+  const objectCode =  `
+  <StyledText
+    key={index}
+    animationProps={{
+      animationname: "${animationType}",
+      duration: "${duration}ms",
+      delay: "${interval}ms",
+      direction: "${direction}",
+      timing: "${timingFunction}",
+      iteration: ${iterationCount},
+      fillmode: "${fillmode}",
+    }}>
+    React-Style-Text
+  </StyledText>
+  `.trim()
+
+  const lettersCode = `
+  const Letters = "React-Style-Text".split("")
+  
+  Letters.map((item, index) =>
+  <StyledText
+    key={index}
+    animationProps={{
+      animationname: "${animationType}",
+      duration: "${duration}ms",
+      delay: "${interval}ms",
+      direction: "${direction}",
+      timing: "${timingFunction}",
+      iteration: ${iterationCount},
+      fillmode: "${fillmode}",
+    }}>
+  {item}
+  </StyledText>)
+    `.trim()
+
+    const typewriterCode = `<StyledText type="typewriter"
+    dataText={[
+    'Sushi',
+    'Pizza',
+    'Brötchen',
+    'Salat'
+  ]} />`.trim()
+
   return (
     <Container>
-      <StyledCode>
-        {
-          animationObject === "object" &&
-          `<MovingComponent
-  type="${animationType}"
-  duration="${duration}ms"
-  delay="${delay}s"
-  direction="${direction}"
-  timing="${timingFunction}"
-  iteration="${iterationCount}"
-  fillMode="${fillMode}">
-  React-Moving-Text
-</MovingComponent>`.trim()
-        }
-
-{
-  animationObject === "letters" &&
-  `Letters.map((letter, index) =>
-      <MovingComponent
-        type="${animationType}"
-        duration="${duration}ms"
-        delay="index * ${interval}ms"
-        direction="${direction}"
-        timing="${timingFunction}"
-        iteration="${iterationCount}"
-        fillMode="${fillMode}">
-        {letter}
-      </MovingComponent>)`.trim()
-
-}
-
-{
-  animationObject === "multiline" &&
-  `<MovingComponent type="typewriter"
-  dataText={[
-  'Sushi',
-  'Pizza',
-  'Brötchen',
-  'Salat'
-]} />`.trim()
-
-}
-
+      <StyledCode onClick={() => {navigator.clipboard.writeText(textAreaRef.current.innerText); swal("Copied!", "Code copied to clipboard!", "success", {button: "Aww yiss!"}) }} ref={textAreaRef}>
+        { animationObject === "object" && objectCode }
+        { animationObject === "letters" && lettersCode }
+        { animationObject === "multiline" &&  typewriterCode}
       </StyledCode>
     </Container>
   )
@@ -67,16 +73,19 @@ const Container = styled.pre`
   display: block;
   position: relative;
   padding: 1rem 2rem;
-  width: 100%;
+  width: 25vw;
   font-size: 1.1rem;
   line-height: 1.5;
   color: ${props => props.theme.colors.white};
+  text-shadow: 0 0 10px rgba(0,0,0,0.5);
   word-break: break-all;
   overflow-wrap: break-word;
   background-color: ${props => props.theme.colors.middle};
   border: 2px solid  ${props => props.theme.colors.primary};
   border-radius: 8px;
-  overflow: auto;
+  overflow: scroll;
+  box-shadow: inset 0 0 10px rgba(0,0,0,0.5);
+  cursor: pointer;
 
   ::selection {
     background-color: ${props => props.theme.colors.primary};
